test(web): ride key-detail e2e on the matrix spec, not a new group (#491)
web-closure failed because the new key-detail flow could not actually run: the
merge gate (`ci-control.yml`) is `pull_request_target` and loads `ci.yml` from
the base branch, so a PR adding a spec to a group's list never runs it — its
pinned claims stay unexecuted and the aggregator refuses the registry as not
closed.

Fix without touching the gate: the key-detail flow now rides `matrix.spec.ts`
(already in a group on main), with the surface's journey, stale-link recovery,
and pinned assertion set (both themes) added as a self-contained describe there.
The spec file's content comes from the PR checkout, so the tests run today.
Revert the inert ci.yml group edit and drop the standalone spec file.

Verified locally: matrix.spec.ts desktop+mobile green, including the key-detail
describe and its Key-declaration pinned set.
